+++ breeze-serialization/src/main/kotlin/MapSerializer.kt
@@ -1,16 +1,20 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-package com.windea.breezeframework.mapper
+package com.windea.breezeframework.serialization
 
 import com.windea.breezeframework.reflect.extensions.*
 
+//TODO 完善
+
 /**
- * 对象的映射器。
+ * 映射的序列化器。
  */
-object ObjectMapper {
-	/**将对象映射为基于可读属性的映射。*/
-	fun <T : Any> map(data: T): Map<String, Any?> {
+object MapSerializer {
+	/**
+	 * 序列化指定对象为映射。
+	 */
+	fun <T : Any> serialize(data: T): Map<String, Any?> {
 		//使用Java反射，映射第一层属性，不进行递归映射
 		return data::class.java.getters.associateBy(
 			{ it.name[3].toLowerCase() + it.name.substring(4) },
@@ -18,13 +22,17 @@ object ObjectMapper {
 		)
 	}
 
-	/**将映射反映射为指定类型的基于可读写属性的对象。*/
-	inline fun <reified T : Any> unmap(map: Map<String, Any?>): T {
-		return unmap(map, T::class.java)
+	/**
+	 * 反序列化指定映射为对象。
+	 */
+	inline fun <reified T : Any> deserialize(map: Map<String, Any?>): T {
+		return deserialize(map, T::class.java)
 	}
 
-	/**将映射反映射为指定类型的基于可读写属性的对象。*/
-	fun <T : Any> unmap(map: Map<String, Any?>, type: Class<T>): T {
+	/**
+	 * 反序列化指定映射为对象。
+	 */
+	fun <T : Any> deserialize(map: Map<String, Any?>, type: Class<T>): T {
 		return runCatching {
 			//存在无参构造时，使用Java反射，直接实例化对象
 			val result = type.getConstructor().newInstance()