+++ CHANGELOG.md
@@ -23,7 +23,9 @@
 * [X] 废弃`breeze-mapper`和`breeze-serializer`并完善`breeze-serialization`
 * [X] 完善`breeze-serialization`
 * [X] 移除`breeze-serializer`
+* [X] breeze-serialization 添加`MapSerializer`
 * [ ] 整理模块`breeze-dsl`中
+* [ ] 
 * [ ] 用于处理CSV和TSV的`Mapper`
 
 ## 1.x.x